Add rule to ignore unused next argument
Express middleware that does not end the request-response cycle must call next() to pass control to the next middleware function.  Error-handler middleware, in particular, always takes four arguments including next().  Since next is being defined as an argument and eslint does not detect its use an unused-vars error will be reported.  To avoid this and to keep from having to litter the project code with eslint-disable-line comments a rule is being added to .eslintrc to ignore the unused "next" argument.